requirements.txt
Mastodon.py==1.7.0
schedule==1.1.0